Tested playTrack
